refactor serverjs into seprate routes
