--- backend/server/routes/database.js
@@ -1,5 +1,0 @@
-const express = require('express');
-const router = express.Router()
-require('dotenv').config();
-
-
